# Remove spinlock in libpas on Linux