### What does this PR do?
Remove spinlock in libpas on Linux

This stress test for `Worker` memory allocation gets about 15% faster



Previous:
-  427 seconds user
 - 323 seconds sys

New:
- 581 seconds user
- 16 seconds sys

Previous:

| Metric | Value | Notes | 
|--------|-------|-------|
| cpu_atom/cycles/ | 1,853,579,678,655 | 
| cpu_core/cycles/ | 2,344,285,658,686 | 
| cpu_atom/instructions/ | 2,326,034,395,212 | 1.25 insn per cycle |
| cpu_core/instructions/ | 3,024,521,916,104 | 1.29 insn per cycle |
| cpu_atom/cache-misses/ | 2,348,429,607 | 44.28% of all cache refs |
| cpu_core/cache-misses/ | 2,898,970,152 | 31.16% of all cache refs |
| cpu_atom/cache-references/ | 5,303,711,115 | 
| cpu_core/cache-references/ | 9,303,752,118 | 
| cpu_atom/branch-misses/ | 2,675,684,043 | 
| cpu_core/branch-misses/ | 2,945,814,093 | 
| context-switches | 153,547,277 | 

New:

| Metric | Value | Notes | 
|--------|-------|-------|
| cpu_atom/cycles/ | 1,798,918,239,365 | 
| cpu_core/cycles/ | 2,193,060,134,318 | 
| cpu_atom/instructions/ | 1,028,765,118,851 | 0.57 insn per cycle |
| cpu_core/instructions/ | 2,499,499,963,385 | 1.14 insn per cycle |
| cpu_atom/cache-misses/ | 1,327,778,729 | 32.85% of all cache refs |
| cpu_core/cache-misses/ | 3,993,164,593 | 22.03% of all cache refs |
| cpu_atom/cache-references/ | 4,042,002,770 | 
| cpu_core/cache-references/ | 18,127,362,686 | 
| cpu_atom/branch-misses/ | 649,190,937 | 
| cpu_core/branch-misses/ | 1,500,244,025 | 
| context-switches | 2,715,342 | 


### How did you verify your code works?

CI

---

```js
// contention-bench.js

const NUM_WORKERS = (navigator.hardwareConcurrency * 0.75) | 0;
const BATCH_SIZE = 10000;
const BATCHES = 100;
import { Worker } from "node:worker_threads";

const workerCode = `
const {parentPort} = require('node:worker_threads');
// Process all batches
globalThis.buffers = [];
for (let batch = 0; batch < ${BATCHES}; batch++) {
    const start = performance.now();
    
    // Create contention by rapidly allocating and deallocating
    for (let i = 0; i < ${BATCH_SIZE}; i++) {
        const buf = new Uint8Array(511);
        for (let j = 0; j < buf.length; j++) {
            buf[j] = j & 0xFF;
        }
        buffers.push(buf);
        
        if (buffers.length > 100) {
            buffers.length = 0;
        }
    }
    
    const end = performance.now();
    parentPort.postMessage({ batch, duration: end - start });
}

parentPort.postMessage('done');
`;

const workers = Array(NUM_WORKERS)
  .fill()
  .map(() => new Worker(workerCode, { type: "module", eval: true }));

let completed = 0;

// Handle each worker completion
workers.forEach((worker, idx) => {
  worker.on("message", ({ done }) => {
    if (done) {
      completed++;
      if (completed === NUM_WORKERS) {
        process.exit(0);
      }
    }
  });
});

```