@robobun 2025-01-03
<div><sup>Updated 2:47 AM PT - Jan 3rd, 2025</sup></div>

@Jarred-Sumner, your commit f5af2ccd412ba52c6eb0399d53800bd6ace0079e is building: [`#8873`](https://buildkite.com/bun/bun/builds/8873)
<!-- generated-comment id=oven-sh/bun#16130 -->